refactored to use omni.kit.viewport.utility
