+++ exts/maticodes.viewport.reticle/config/extension.toml
@@ -1,5 +1,5 @@
 [package]
-version = "1.2.0"
+version = "1.3.0"
 title = "Camera Reticle"
 description="Adds a camera reticle featuring composition guidelines, safe area guidelines, and letterbox."
 authors=["Matias Codesal <mati@matiascodesal.com>"]
@@ -15,6 +15,7 @@ icon = "icons/icon.png"
 # Use omni.ui to build simple UI
 [dependencies]
 "omni.ui.scene" = {}
+"omni.kit.viewport.utility" = {}
 
 # Main python module this extension provides, it will be publicly available as "import omni.hello.world".
 [[python.module]]
